+++ .editorconfig
@@ -18,6 +18,10 @@ indent_size = 4
 [*.yml]
 indent_size = 2
 
-[*.{ts, js}]
+[*.ts]
+indent_size = 4
+max_line_length = 119
+
+[*.js]
 indent_size = 4
 max_line_length = 119
