+++ .gitignore
@@ -5,4 +5,9 @@ __pycache__
 .vscode
 /venv/
 /build/*
+*/node_modules/
 /node_modules/
+.yarn
+yarn.lock
+yarn-error.log
+.yarnrc.yml
