+++ js/designer_components/index.ts
@@ -1,0 +1,5 @@
+export { DesignerMultiSelectDropDown } from "./DesignerMultSelectDropDown";
+export { DesignerQuill } from "./DesignerQuill";
+export { DesignerSlider } from "./DesignerSlider";
+export { DesignerSwitch } from "./DesignerSwitch";
+export { DesignerTabs } from "./DesignerTabs";
